Remove data store
