--- assets/js/data/incompatible-extensions/action-types.ts
@@ -1,4 +1,0 @@
-export enum ACTION_TYPES {
-	ADD_INCOMPATIBLE_EXTENSION = 'ADD_INCOMPATIBLE_EXTENSION',
-	REMOVE_INCOMPATIBLE_EXTENSION = 'REMOVE_INCOMPATIBLE_EXTENSION',
-}
