--- assets/js/data/incompatible-extensions/actions.ts
@@ -1,18 +1,0 @@
-/**
- * Internal dependencies
- */
-import { ACTION_TYPES } from './action-types';
-
-export const addIncompatibleExtension = ( extension: string ) => {
-	return {
-		type: ACTION_TYPES.ADD_INCOMPATIBLE_EXTENSION,
-		extension,
-	};
-};
-
-export const removeIncompatibleExtension = ( extension: string ) => {
-	return {
-		type: ACTION_TYPES.REMOVE_INCOMPATIBLE_EXTENSION,
-		extension,
-	};
-};
