--- assets/js/data/incompatible-extensions/default-state.ts
@@ -1,7 +1,0 @@
-export interface IncompatibleExtensionsState {
-	extensions: string[];
-}
-
-export const defaultIncompatibleExtensionsState: IncompatibleExtensionsState = {
-	extensions: [],
-};
