--- assets/js/data/incompatible-extensions/index.ts
@@ -1,34 +1,0 @@
-/**
- * External dependencies
- */
-import { createReduxStore, register } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
-import * as actions from './actions';
-import * as selectors from './selectors';
-import reducer from './reducers';
-import { DispatchFromMap, SelectFromMap } from '../mapped-types';
-
-const STORE_KEY = 'wc/store/incompatible-extensions';
-const config = {
-	reducer,
-	actions,
-	selectors,
-};
-const store = createReduxStore( STORE_KEY, config );
-register( store );
-
-export const INCOMPATIBE_EXTENSIONS_STORE_KEY = STORE_KEY;
-
-declare module '@wordpress/data' {
-	function dispatch(
-		key: typeof STORE_KEY
-	): DispatchFromMap< typeof actions >;
-	function select( key: typeof STORE_KEY ): SelectFromMap<
-		typeof selectors
-	> & {
-		hasFinishedResolution: ( selector: string ) => boolean;
-	};
-}
