+++ assets/js/editor-components/incompatible-extension-notice/use-incompatible-extensions-notice.ts
@@ -1,21 +1,27 @@
 /**
  * External dependencies
  */
-import { useSelect } from '@wordpress/data';
+import { getSetting } from '@woocommerce/settings';
 
 export const useIncompatibleExtensionNotice = (): [
 	{ [ k: string ]: string } | null,
 	string[],
 	number
 ] => {
-	const { incompatibleExtensions } = useSelect( ( select ) => {
-		const { getIncompatibleExtensions } = select(
-			'wc/store/incompatible-extensions'
-		);
-		return {
-			incompatibleExtensions: getIncompatibleExtensions(),
-		};
-	}, [] );
+	interface GlobalIncompatibleExtensions {
+		id: string;
+		title: string;
+	}
+
+	const incompatibleExtensions: Record< string, string > = {};
+
+	if ( getSetting( 'incompatibleExtensions' ) ) {
+		getSetting< GlobalIncompatibleExtensions[] >(
+			'incompatibleExtensions'
+		).forEach( ( extension ) => {
+			incompatibleExtensions[ extension.id ] = extension.title;
+		} );
+	}
 
 	const incompatibleExtensionSlugs = Object.keys( incompatibleExtensions );
 	const incompatibleExtensionCount = incompatibleExtensionSlugs.length;
